# doc/openapi.yaml
openapi: 3.1.0
info:
  title: API REST de Notium
  version: 1.0.0
  summary: Backend de sincronización para la plataforma Notium offline-first.
  description: |
    API REST del backend de **Notium**, la plataforma de toma de notas
    offline-first (programa ADSO — SENA). Implementada con Node.js + Express +
    PostgreSQL (sección 6.2 y ADR-05 del documento de arquitectura).

    El núcleo del sistema es la sincronización por lotes (`/sync/push` y
    `/sync/pull`), definida en la sección 5.5 del documento de arquitectura:

    - Los identificadores (`uuid`) los genera **el cliente** (UUID v4); el
      servidor solo los acepta (sección 5.1).
    - `POST /sync/push` es **idempotente por la clave `(uuid, version)`**:
      reenviar la misma operación tras un reintento no duplica datos.
    - Los conflictos se resuelven con **Last-Write-Wins** sobre `updated_at`,
      usando `version` para detectar la edición concurrente (sección 5.3).
    - Las eliminaciones se propagan como **tombstones** (`is_deleted = true`,
      sección 5.4).
    - Toda comunicación usa HTTPS/TLS; la autenticación es JWT
      (access + refresh) y las contraseñas se almacenan con hash bcrypt.
  contact:
    name: Jhonatan Castro
    email: shonano247@gmail.com
  license:
    name: Uso académico — SENA (ADSO)
    identifier: LicenseRef-Academico-SENA
servers:
  - url: https://api.notium.example.com/v1
    description: Servidor de producción (infraestructura de bajo costo, sección 2.3).
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Autenticación]
      operationId: iniciarSesion
      summary: Iniciar sesión
      description: |
        Valida las credenciales y devuelve el par de tokens JWT
        (access + refresh). El cliente los almacena en `flutter_secure_storage`
        para habilitar la validación local de sesión estando offline (sección 8).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sesión iniciada correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SesionResponse'
        '400':
          $ref: '#/components/responses/ErrorValidacion'
        '401':
          description: Credenciales inválidas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /auth/logout:
    post:
      tags: [Autenticación]
      operationId: cerrarSesion
      summary: Cerrar sesión
      description: |
        Invalida el refresh token del dispositivo en el servidor. Soporta la
        mitigación de la sección 9: ante pérdida o robo del dispositivo, la
        sesión puede invalidarse remotamente y surtirá efecto en la próxima
        conexión.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Sesión cerrada; el refresh token queda invalidado.
        '401':
          $ref: '#/components/responses/NoAutorizado'
...
  /attachments:
    post:
      tags: [Adjuntos]
      operationId: subirAdjunto
      summary: Subir un archivo adjunto
      description: |
        Carga el archivo físico de un adjunto mediante `multipart/form-data`.
        Los adjuntos viajan **fuera** del lote de `/sync/push`: el lote solo
        sincroniza los metadatos de la entidad `ADJUNTO`; el binario se sube por
        este endpoint. El tamaño máximo por archivo es **10 MB** (RNF-05); un
        archivo mayor se rechaza con `413` y el cliente marca el registro con
        `sync_status = ERROR` (no se reintenta).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [uuid, nota_uuid, archivo]
              properties:
                uuid:
                  type: string
                  format: uuid
                  description: Identificador del adjunto, generado en el cliente (UUID v4).
                nota_uuid:
                  type: string
                  format: uuid
                  description: Nota a la que pertenece el adjunto.
                device_id:
                  type: string
                  description: Dispositivo que origina la subida, para trazabilidad.
                archivo:
                  type: string
                  format: binary
                  description: Contenido binario del archivo (máximo 10 MB, RNF-05).
      responses:
        '201':
          description: Adjunto almacenado. Devuelve los metadatos con la `url_remota` asignada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adjunto'
        '400':
          $ref: '#/components/responses/ErrorValidacion'
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '404':
          description: La nota referenciada no existe o no pertenece al usuario autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Ya existe un adjunto con ese `uuid` y contenido distinto (violación de idempotencia).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: El archivo excede el límite de 10 MB por adjunto (RNF-05).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /attachments/{uuid}:
    parameters:
      - $ref: '#/components/parameters/UuidAdjunto'
    get:
      tags: [Adjuntos]
      operationId: descargarAdjunto
      summary: Descargar un archivo adjunto
      description: |
        Devuelve el contenido binario del adjunto. El servidor valida que el
        adjunto pertenezca al usuario autenticado (control de acceso por
        `usuario_uuid`, sección 8).
      responses:
        '200':
          description: Contenido binario del archivo.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '410':
          description: El adjunto fue eliminado (tombstone) y su binario ya no está disponible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Adjuntos]
      operationId: eliminarAdjunto
      summary: Eliminar un adjunto (soft delete)
      description: |
        Marca el adjunto como eliminado (`is_deleted = true`, tombstone de la
        sección 5.4) para propagar la eliminación a los demás dispositivos vía
        `/sync/pull`. La purga física del binario ocurre después de la ventana
        de gracia, igual que la de los tombstones locales.
      responses:
        '204':
          description: Adjunto marcado como eliminado.
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '404':
          $ref: '#/components/responses/NoEncontrado'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token JWT emitido por `/auth/login` o `/auth/refresh`.
        El cliente lo conserva en `flutter_secure_storage` (sección 8).
  parameters:
    UuidAdjunto:
      name: uuid
      in: path
      required: true
      description: Identificador (UUID v4) del adjunto, generado en el cliente.
      schema:
        type: string
        format: uuid
  responses:
    NoAutorizado:
      description: Token ausente, inválido o expirado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NoEncontrado:
      description: El recurso no existe o no pertenece al usuario autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ErrorValidacion:
      description: La solicitud no cumple el esquema o contiene datos inválidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    SyncStatus:
      type: string
      description: Estado de sincronización de un registro (sección 4.2). Metadato gestionado por el cliente.
      enum: [PENDING, SYNCED, CONFLICT, ERROR]
...
    Usuario:
      type: object
      description: Entidad USUARIO (sección 4.1). No incluye tokens de sesión; estos viven en `flutter_secure_storage` del cliente.
      required: [uuid, nombre, email]
      properties:
        uuid:
          type: string
          format: uuid
          description: Identificador único generado en el cliente (UUID v4, sección 5.1).
        nombre:
          type: string
          description: Nombre completo del usuario.
        email:
          type: string
          format: email
          description: Correo electrónico, único por cuenta.
...
    Adjunto:
      type: object
      description: Entidad ADJUNTO (sección 4.1). El binario viaja por `/attachments`, no dentro del lote de sincronización.
      required: [uuid, nota_uuid]
      properties:
        uuid:
          type: string
          format: uuid
          description: Identificador único generado en el cliente (UUID v4).
        nota_uuid:
          type: string
          format: uuid
          description: Nota a la que pertenece el adjunto.
        ruta_local:
          type: [string, 'null']
          description: Ruta del archivo en el dispositivo. Metadato local del cliente; el servidor la conserva solo como dato informado.
        url_remota:
          type: [string, 'null']
          format: uri
          description: URL de descarga asignada por el servidor tras la subida.
        sync_status:
          $ref: '#/components/schemas/SyncStatus'
          description: Metadato local del cliente.
        is_deleted:
          type: boolean
          description: Marca de borrado lógico (tombstone).
        version:
          type: integer
          description: Contador incremental para detectar ediciones concurrentes.
        updated_at:
          type: string
          format: date-time
          description: Última modificación del registro de metadatos.
...
    LoginRequest:
      type: object
      description: Credenciales de inicio de sesión.
      required: [email, contrasena]
      properties:
        email:
          type: string
          format: email
          description: Correo electrónico de la cuenta.
        contrasena:
          type: string
          format: password
          description: Contraseña de la cuenta.
    RefreshRequest:
      type: object
      description: Solicitud de refresco o invalidación de sesión.
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Refresh token vigente emitido en el login.
    TokensResponse:
      type: object
      description: Par de tokens JWT (sección 8). El cliente los almacena en `flutter_secure_storage`.
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: Token de acceso JWT de corta duración.
        refresh_token:
          type: string
          description: Token de refresco de larga duración.
        token_type:
          type: string
          description: Tipo de token; siempre `Bearer`.
          const: Bearer
        expires_in:
          type: integer
          description: Segundos de vigencia del access token; base de la validación local de expiración estando offline.
    SesionResponse:
      type: object
      description: Sesión completa devuelta al registrar o iniciar sesión.
      required: [usuario, tokens]
      properties:
        usuario:
          $ref: '#/components/schemas/Usuario'
        tokens:
          $ref: '#/components/schemas/TokensResponse'

    # ------------------------------------------------------------------
    # Utilidades
    # ------------------------------------------------------------------
    Error:
      type: object
      description: Esquema de error estándar de la API.
      required: [codigo, mensaje]
      properties:
        codigo:
          type: string
          description: Código de error estable, apto para lógica del cliente (por ejemplo, `EMAIL_DUPLICADO`, `ADJUNTO_EXCEDE_LIMITE`, `TOKEN_EXPIRADO`).
        mensaje:
          type: string
          description: Descripción legible del error, en español.
        detalle:
          type: [string, 'null']
          description: Información adicional de diagnóstico (campo inválido, límites, etc.).
...